Move generic datasource tests to test_file_based_datasource
- Move test_read_ray_remote_args from test_text.py to test_file_based_datasource.py
- Move test_fsspec_http_file_system from test_text.py to test_file_based_datasource.py
- Remove test_read_tfrecords_ray_remote_args from test_tfrecords.py
- Remove test_write_min_rows_per_file from test_tfrecords.py

These tests cover generic FileBasedDatasource behavior, not format-specific functionality.
